+++ HISTORY.md
@@ -7,6 +7,7 @@ unreleased
     - Improves compatibility with Node.js 8 nightly
   * Send complete HTML document in redirect & error responses
   * Set default CSP header in redirect & error responses
+  * Use `res.getHeaderNames()` when available
   * deps: debug@2.6.1
     - Allow colors in workers
     - Deprecated `DEBUG_FD` environment variable set to `3` or higher